@@ -1,7 +1,7 @@
 cmake_minimum_required(VERSION 3.12)
 project(visteontraining VERSION 1.0)
 
-set(CMAKE_CXX_STANDARD 14)
+set(CMAKE_CXX_STANDARD 17)
 set(CMAKE_CXX_STANDARD_REQUIRED True)
 
 add_executable(engine
